sf-ai-agentforce-testing-cirra-ai/validation/requirements.txt
pytest>=8.0.0
pytest-cov>=4.1.0
pyyaml>=6.0
jsonschema>=4.21.0
rich>=13.0.0
tabulate>=0.9.0
